# Fixed type's for session objects to include `CachedSession`. - Fixed logic bug in the `node` function. - Changed `Item.patch` type from `int` to `float`.

+++ async_garlandtools/_types.py
@@ -260,7 +260,7 @@ class Item(TypedDict):
     "possibly use as a bool"
     elvl: NotRequired[int]
     jobs: NotRequired[int]
-    patch: int
+    patch: float
     patchCategory: int
     price: int
     ilvl: int
